+++ tests/Pest.php
@@ -13,7 +13,7 @@ use Flobbos\PageComposer\Tests\TestCase;
 |
 */
 
-uses(TestCase::class)->in('Feature');
+uses(TestCase::class)->in('Feature', 'Unit');
 
 /*
 |--------------------------------------------------------------------------